requirements.txt
streamlit>=1.38.0
langgraph>=0.2.0
langchain-core>=0.3.0
gspread>=6.0.0
google-auth>=2.29.0
litellm>=1.40.0
pandas>=2.2.0
nest_asyncio>=1.6.0
